# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The `\"detailed\"` error envelope every handler renders, and what OpenAPI advertises.\n\n`details` is present only on a 422, where it holds one entry per offending field.",
        "properties": {
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field validation failures. Present only on a 422.",
            "title": "Details"
          },
          "error_code": {
            "description": "A stable machine readable code, such as `NOT_FOUND`.",
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "description": "A human readable summary safe to surface to a caller.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "description": "The request id, echoed in the `X-Request-ID` header.",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "The HTTP status code, repeated in the body.",
            "title": "Status",
            "type": "integer"
          },
          "success": {
            "default": false,
            "description": "Always `false` on an error response.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "status",
          "message",
          "request_id",
          "error_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ValidationErrorDetail": {
        "description": "One offending field in a 422, carrying its location and reason but never its input.",
        "properties": {
          "field": {
            "description": "Dotted path to the offending field, or `_root` for the body itself.",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Why the value was rejected.",
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "The pydantic error type, such as `missing` or `string_type`.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ValidationErrorDetail",
        "type": "object"
      },
...
      "WorkspaceCreate": {
        "description": "A new workspace. The name is unique across the environment.",
        "properties": {
          "auto_apply": {
            "default": false,
            "title": "Auto Apply",
            "type": "boolean"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "engine": {
            "default": "terraform",
            "enum": [
              "terraform",
              "tofu"
            ],
            "title": "Engine",
            "type": "string"
          },
          "engine_version": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Engine Version",
            "type": "string"
          },
          "file_triggers_enabled": {
            "default": true,
            "title": "File Triggers Enabled",
            "type": "boolean"
          },
          "name": {
            "maxLength": 90,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Name",
            "type": "string"
          },
          "plan_assume_role_arns": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Plan Assume Role Arns",
            "type": "array"
          },
          "plan_role_arn": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Role Arn"
          },
          "plan_secret_arns": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Plan Secret Arns",
            "type": "array"
          },
          "project_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^prj-(default|[0-9A-HJKMNP-TV-Z]{26})$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "run_role_arn": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Role Arn"
          },
          "speculative_plans": {
            "default": true,
            "title": "Speculative Plans",
            "type": "boolean"
          },
          "tracked_branch": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracked Branch"
          },
          "trigger_patterns": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Trigger Patterns",
            "type": "array"
          },
          "vcs_repo": {
            "anyOf": [
              {
                "maxLength": 140,
                "pattern": "^[A-Za-z0-9-]+/[A-Za-z0-9._-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vcs Repo"
          },
          "working_directory": {
            "default": "",
            "title": "Working Directory",
            "type": "string"
          }
        },
        "required": [
          "engine_version",
          "name"
        ],
        "title": "WorkspaceCreate",
        "type": "object"
      },
      "WorkspaceList": {
        "description": "Every workspace, newest last by id, which is a ULID and so time ordered.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceListItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "WorkspaceList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Workspaces, variables, config versions and runs, as the gateway routes them.",
    "title": "WebbPulse Terraform control plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "get": {
        "description": "Workspaces in this environment, each with its run role setup and newest run.\n\nThe filters and the sort compose, so a project's workspaces can be searched and\nordered in one request. A `project_id` naming a project that holds nothing, or none\nat all, is an empty list. A run's API token sees only its own workspace.",
        "operationId": "list_workspaces_api_v1_workspaces_get",
        "parameters": [
          {
            "description": "Only the workspaces in this project. `prj-default` is the default project.",
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "pattern": "^prj-(default|[0-9A-HJKMNP-TV-Z]{26})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only the workspaces in this project. `prj-default` is the default project.",
              "title": "Project Id"
            }
          },
          {
            "description": "Only the workspaces whose name contains this, ignoring case.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 90,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only the workspaces whose name contains this, ignoring case.",
              "title": "Search"
            }
          },
          {
            "description": "`name` or `-name`, `-latest_run` (newest run first, never run last), `-updated_at` (latest change first), `-created_at`, or `status` (runs that need attention first). Without it the list is oldest first.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "name",
                    "-name",
                    "-latest_run",
                    "-updated_at",
                    "-created_at",
                    "status"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`name` or `-name`, `-latest_run` (newest run first, never run last), `-updated_at` (latest change first), `-created_at`, or `status` (runs that need attention first). Without it the list is oldest first.",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "List Workspaces",
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "description": "Create a workspace. The name has to be free.\n\nThe run role is optional here on purpose: its trust policy names the workspace\nid as the external id, so the role cannot exist until the workspace does. The\nresponse's `run_role_setup` carries everything needed to build it, and\n`PATCH /workspaces/{id}` attaches it afterwards.\n\nA `vcs_repo` is resolved through the environment's GitHub App: the id, the\ninstallation and the canonical name are recorded, and `tracked_branch` defaults to\nthe repository's default branch. A repository the App is not installed on is a 422\ncarrying `VCS_REPO_NOT_INSTALLED`, as is a `project_id` naming no project, carrying\n`PROJECT_NOT_FOUND`.\n\nCreating a workspace with `auto_apply` on lets anyone who can start a run apply\nit, so it takes `admin` and the step-up, as turning it on later does. Naming a\nplan role, reader roles or plan secrets takes the same, as setting them later does,\nand naming a run role takes the step-up.",
        "operationId": "create_workspace_api_v1_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Create Workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}": {
      "delete": {
        "description": "Delete one workspace with its finished runs, current state and variables.\n\nA safe delete is a 409 carrying `WORKSPACE_MANAGES_RESOURCES` while the current\nstate tracks any resource instance: queue a destroy plan and apply it first, or\npass `force=true` to delete anyway and leave those resources unmanaged. Either\nmode is a 409 carrying `WORKSPACE_HAS_ACTIVE_RUN` while a run on the workspace\nhas not finished. Nothing is removed on a refusal.",
        "operationId": "delete_workspace_api_v1_workspaces__workspace_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^ws-[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "description": "Skip the managed resources check and delete even though state still tracks resources.",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "description": "Skip the managed resources check and delete even though state still tracks resources.",
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Delete Workspace",
        "tags": [
          "workspaces"
        ]
      },
      "get": {
        "description": "One workspace by id, with its run role setup.",
        "operationId": "get_workspace_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^ws-[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "description": "Edit one workspace. The name and the id are not editable.\n\nThe body is JSON Merge Patch: an omitted key leaves the stored value exactly\nas it was, and an explicit null on any field but `engine` and `engine_version`\nclears it. `model_dump(exclude_unset=True)` is what keeps the two apart, so a\nfield is only touched when the request carried its key.\n\nChanging or clearing `run_role_arn` drops the recorded check outcome, so the\nrole reads as unchecked until `run-role/check` says otherwise.\n\nConnecting another `vcs_repo` resolves it through the GitHub App the way the\ncreate does, and a null disconnects the repository.\n\nChanging a run role, the plan role or the reader roles is the workspace's AWS\nconnection, so a person has to have signed in within the step-up window; resending\nthe stored value is not a change.\n\nChanging `plan_role_arn`, `plan_assume_role_arns` or `plan_secret_arns` decides what\nevery plan, a pull request plan included, may reach, so it takes `admin`, or a run\ntoken holding the factory grant, as well as the step-up. The reader roles and the\nplan secrets together must fit the plan's session policy, or the edit is a 422\n`PLAN_SESSION_POLICY_TOO_LARGE`.\n\nChanging `run_api_token_scopes` hands every later run of this workspace a key on\nthis API, so it takes `admin` as well as the step-up.\n\nChanging `auto_apply` turns `runs:write` into the power to apply, so it takes\n`admin` and is recorded under `workspaces.workspace.auto_apply`. Turning it on also\ntakes the step-up, since it hands every later run's confirmation to the system.\n\n`global_remote_state` and `remote_state_consumer_ids` decide which other workspaces'\nruns may read this workspace's non-sensitive outputs. Every named consumer has to\nexist, or the edit is a 422 `REMOTE_STATE_CONSUMER_NOT_FOUND`.",
        "operationId": "update_workspace_api_v1_workspaces__workspace_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^ws-[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Update Workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
...
  }
}